Refactor numba lapack codegen
